Remove old REINFORCE
